+++ src/browserify.index.js
@@ -9,6 +9,5 @@ import RxDB from './index.js';
 
 RxDB.plugin(require('pouchdb-adapter-idb'));
 RxDB.plugin(require('pouchdb-adapter-http'));
-RxDB.plugin(require('pouchdb-replication'));
 
 window['RxDB'] = RxDB;